requirements.txt
requests
pytest
allure-pytest
jsonschema
